[IMP] mail_broker_telegram: Pass logic to component
