+++ mail_broker_telegram/__init__.py
@@ -1,1 +1,2 @@
 from . import models
+from . import services
