+++ mail_broker_telegram/models/__init__.py
@@ -1,3 +1,2 @@
 from . import mail_broker_channel
-from . import mail_message_broker
 from . import mail_broker
