--- mail_broker_telegram/tests/__init__.py
@@ -1,2 +1,0 @@
-from . import test_mail_broker
-from . import test_mail_broker_controller
